trackexec: improve and refactor
This commit mainly brings:

* Enhanced idle period concept -> pt sections:
  - signaling an instruction as an idle period in the code caused
    trackexec to record timestamps around this instruction, for more
    precise overall timestamping.
  - this commit extends this concept by allowing several instructions in
    the "with" context, and rename this feature to "precise timestamping"
    sections instead of "idle" periods, since developers sometimes do not
    know exactly which exact inctruction is time consuming at first.

* Support for monitoring a single module (instead of a package)

* Fix sections with precise timestamping

* Numba-accelerated code is now compiled when the module is loaded

* A command line tool walt-server-trackexec-analyse allows
  to list "hot points" in the source code. This required to change
  the block on-disk format: record both the file_id and the line
  number for each stack item.

* There could be issues getting non-monotonic timestamps (probably
  caused by NTP resyncs). This commit fixes this by ensuring any
  recorded timestamp is higher or equal than the previous one.

* The code has been much refactored.

+++ server/walt/server/trackexec/const.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 # MAP_BLOCK_UINT16_SIZE and MAP_BLOCK_ID_SHIFT were selected
 # according to tests giving high compressability of the resulting
 # MAP_FILE_SIZE chunks.
@@ -16,8 +18,8 @@ DAY_AS_TS = 24 * HOUR_AS_TS
 
 
 class OpCodes:
-    CALL = 65535
-    RETURN = 65534
-    TIMESTAMP = 65533
-    END = 0  # end of block
+    CALL = np.uint16(65535)
+    RETURN = np.uint16(65534)
+    TIMESTAMP = np.uint16(65533)
+    END = np.uint16(0)  # end of block
     # other opcodes are line numbers
